synchronous AES encryption using crypto-js
This is supposed to work both in the browser
and in Node. Should find one of the variants
that works asynchronously eventually, to
avoid browser lockup.

We can't put the entire IV into our message,
or the shortest possible message becomes
256 bits. 225 bits is an important target
for us. The strategy chosen here is to
reduce the IV to 32 bits and use a hash to
"expand" it (which of course doesn't give us
any more entropy).