increase IV seed size to 96 bits
This is very close to as much as we can get
into a 225-bit target (224 bits are used!).
Still exactly one block of data fits -- no
more is possible with AES128 in 225 bits
in any case.

IV size overhead matters less and less as
the message size increases, so we might as
well take the additional security.

IV entropy should probably scale with target
size -- there is no need to implement this
inelegance for our 1683-bit target.