functions for adding and stripping padding
Padding is essential to avoid certain sequences
never getting used. E.g. if we only use 100 bits
out of 225 possible bits, then if we leave the
upper 125 bits at zero, then that run of highly
nonrandom data might be detectable in the output
sequences.

The correct solution is to add 125 bits of
random padding, and then take the lower-order
100 bits when decoding. (For this reason it's
also necessary to have a set message length
depending on the container length, which
basically means padding out the message until
it is as large as any message that could fit.
That way we know how many lower-order bits to
keep.)

(We're assuming the 100 data bits are already
indistinguishable from random bits without the
key -- the encryption should ensure this.)