full conversion from bits to sequences and back
Parts not integrated yet because they could be
configurable and could be done in many different
ways:
    - Encoding
    - Encryption

Encoding is nice-to-have to save space. For
English text, some sort of 6-bit character set
would be ideal, or possibly Huffman coding.
There is no need for preserving case, but it is
important to include both letters and numbers,
which already puts us at a bare minimum of 36
characters.

Encryption is NOT optional as otherwise the data
(and thus the sequence) may appear non-random,
e.g. long ordered runs.