setting up build system for frontend
Using browserify and Google Closure. We want
to be able to deploy in several different
configurations (minified/not minified and
all-in-one-file/normal), so we employ a
Python build script to do all of this for
us.

Currently only producing the all-in-one versions
(and the frontend doesn't do anything useful
yet, merely gives an indication that it's
been loaded).

The all-in-one version is intended to be easier
to use offline.