rework deps to git submodules. include jquery-ui
jQuery UI "sortable" seems perfect for this
application.

The application is now enormous (getting close
to a megabyte before minification). This is
mostly because we include _all_ of jQuery UI.

Once the project UI has stabilized, working out
which parts of jQuery UI we actually need and
omitting the rest is an absolute must.